+++ src/server/interfaces/index.ts
@@ -1,2 +1,5 @@
 export * from './common';
 export * from './file';
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface ServerApplicationState {}